fix: 현재 윈도우 너비에 따라 알맞는 헤더 하나만 표시하도록
