+++ src/App.jsx
@@ -6,7 +6,6 @@ import BookDetail from "./component/book/BookDetail";
 import Footer from "./component/utils/Footer";
 import NotFound from "./component/utils/NotFound";
 import Header from "./component/utils/Header";
-import MobileHeader from "./component/utils/MobileHeader";
 import Information from "./component/information/Information";
 import Main from "./component/main/Main";
 import Search from "./component/search/Search";
@@ -52,7 +51,6 @@ function App() {
       <div id="portal" />
       <Portals />
       <Header />
-      <MobileHeader />
       <Routes>
         <Route path="/" element={<Main />} />
         <Route path="/41" element={<ELibraryIn42Box />} />
